+++ include/vinecopulib/bicop/joe.hpp
@@ -36,14 +36,10 @@ private:
     const Eigen::Ref<const Eigen::VectorXd>& parameters);
 
   // pdf
-  Eigen::VectorXd pdf_raw(const Eigen::MatrixXd& u);
-
   Eigen::VectorXd pdf_raw(const Eigen::MatrixXd& u,
                           const Eigen::MatrixXd& parameters);
 
   // inverse hfunction
-  Eigen::VectorXd hinv1_raw(const Eigen::MatrixXd& u);
-
   Eigen::VectorXd hinv1_raw(const Eigen::MatrixXd& u,
                             const Eigen::MatrixXd& parameters);